refactor(avatar): refactor to top level package
