+++ .changeset/nervous-keys-wink.md
@@ -1,0 +1,6 @@
+---
+'@project44-manifest/react-avatar': minor
+'@project44-manifest/react': minor
+---
+
+Refactor avatar out of react package to its own top level package
