+++ packages/avatar/src/index.ts
@@ -1,0 +1,2 @@
+export * from './Avatar';
+export * from './Avatar.types';
