+++ packages/react/package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "@internationalized/date": "^3.0.0",
+    "@project44-manifest/react-avatar": "^0.0.0",
     "@project44-manifest/react-button": "^1.0.2",
     "@project44-manifest/react-css-baseline": "^0.1.0",
     "@project44-manifest/react-provider": "^0.1.0",
