--- packages/react/src/components/Avatar/index.ts
@@ -1,1 +1,0 @@
-export * from './Avatar';
